quality: fix runtime-backed daemon routing for PR reconcile + review cleanups
Correctness (load-bearing):
- Expose reconcileOnFocus + syncLanePr through the daemon action allowlist
  (registry.ts). Without this both new PR actions were rejected in the
  runtime-backed production build (manual ⟳ a silent no-op, post-auth
  reconcileNow an unhandled rejection).
- Route reconcile-on-focus to the daemon in production: scheduleReconcile now
  takes a runner that re-resolves the live context at fire time and calls the
  daemon pr.reconcileOnFocus action when the desktop context is dormant
  (prService null). Fixes both the dead-in-production hook and the stale-closure
  captured-prService concern.
- Guard the fire-and-forget reconcileNow with .catch.

Cleanups:
- Delete dead statusView (~90 lines) + isGitHubAppRepoAccessPending alias; move
  its test to githubIntegrationStatus.test.ts.
- De-duplicate ingress-prune SQL into pruneIngressEventRowsForProject
  (dbMaintenanceApi.ts), called from automationService + kvDb maintenance.
- Move describeGithubCliBanner into the shared githubIntegrationStatus helper.
- Drop dead ?? "global" dismiss-key fallback.
- Fix stuck ⟳ spinner: split the pr-reconcile subscription into its own effect
  so a linkedPr change can't strand the idle-hide timer.
